Move query insensitivity & special chars ignoring tests out
