+++ tests/search/vespa/test_vespasearch.py
@@ -1,5 +1,3 @@
-import time
-
 import pytest
 from db_client.models.dfce.family import FamilyDocument
 from sqlalchemy import update
@@ -207,48 +205,3 @@ def test_search_with_deleted_docs(test_vespa, monkeypatch, data_client, data_db)
     all_deleted_count = len(all_deleted_body["families"])
     assert start_family_count > one_deleted_count > all_deleted_count
     assert len(all_deleted_body["families"]) == 0
-
-
-@pytest.mark.search
-def test_case_insensitivity(test_vespa, data_db, monkeypatch, data_client):
-    monkeypatch.setattr(search, "_VESPA_CONNECTION", test_vespa)
-    _populate_db_families(data_db)
-
-    lower_body = _make_search_request(data_client, {"query_string": "the"})
-    upper_body = _make_search_request(data_client, {"query_string": "THE"})
-
-    assert lower_body["families"] == upper_body["families"]
-
-
-@pytest.mark.search
-def test_punctuation_ignored(test_vespa, data_db, monkeypatch, data_client):
-    monkeypatch.setattr(search, "_VESPA_CONNECTION", test_vespa)
-    _populate_db_families(data_db)
-
-    regular_body = _make_search_request(data_client, {"query_string": "the"})
-    punc_body = _make_search_request(data_client, {"query_string": ", the."})
-    accent_body = _make_search_request(data_client, {"query_string": "thë"})
-
-    assert (
-        sorted([f["family_slug"] for f in punc_body["families"]])
-        == sorted([f["family_slug"] for f in regular_body["families"]])
-        == sorted([f["family_slug"] for f in accent_body["families"]])
-    )
-
-
-@pytest.mark.search
-def test_accents_ignored(
-    test_vespa,
-    data_db,
-    monkeypatch,
-    data_client,
-):
-    monkeypatch.setattr(search, "_VESPA_CONNECTION", test_vespa)
-    _populate_db_families(data_db)
-
-    start = time.time()
-    body = _make_search_request(data_client, {"query_string": "the"})
-    end = time.time()
-
-    request_time_ms = 1000 * (end - start)
-    assert 0 < body["query_time_ms"] < body["total_time_ms"] < request_time_ms
